Move DeviceType to the types module
