+++ rust/agama-dbus-server/src/network/error.rs
@@ -18,8 +18,6 @@ pub enum NetworkStateError {
     InvalidWirelessMode(String),
     #[error("Connection '{0}' already exists")]
     ConnectionExists(Uuid),
-    #[error("Invalid device type: '{0}'")]
-    InvalidDeviceType(u8),
     #[error("Invalid security wireless protocol: '{0}'")]
     InvalidSecurityProtocol(String),
     #[error("Adapter error: '{0}'")]
