+++ rust/agama-dbus-server/src/network/nm/model.rs
@@ -7,9 +7,10 @@
 /// Using the newtype pattern around an String is enough. For proper support, we might replace this
 /// struct with an enum.
 use crate::network::{
-    model::{DeviceType, IpMethod, SecurityProtocol, WirelessMode},
+    model::{IpMethod, SecurityProtocol, WirelessMode},
     nm::error::NmError,
 };
+use agama_lib::network::types::DeviceType;
 use std::fmt;
 
 #[derive(Debug, PartialEq)]
